backend/requirements.txt
django==2.1.5
djangorestframework==3.9.0
psycopg2==2.7.6.1
django-environ==0.4.5
djangorestframework-simplejwt==3.3
